requirements.txt
discord.py==2.3.2
pymongo==4.6.1
dnspython==2.5.0
Flask==3.0.0
PyNaCl==1.5.0